[pontoon] New study for Pontoon to obtain the translation status
This adds a new study to obtain the latest translation
status for each entity-locale pair. It fetches the most
recent user actions and determines the current status
of the translation.

The new study is called `enrich_latest_translation_status` and
can be configured with the `out_index` and the `alias`. The default
alias is `pontoon_latest_translation_status`.

Signed-off-by: Jose Javier Merchante <jjmerchante@bitergia.com>